minor fix
fixed two audio listeners in start/end screens, moved restartbutton.cs into endscreenmanagement.cs
