+++ Bee Balloon Game/Assets/Scenes/StartScene.unity
@@ -763,7 +763,7 @@ AudioListener:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1956513725}
-  m_Enabled: 1
+  m_Enabled: 0
 --- !u!20 &1956513727
 Camera:
   m_ObjectHideFlags: 0
