--- Bee Balloon Game/Assets/Scripts/EndScreenManager.cs
@@ -1,15 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI;
-
-public class EndScreenManager : MonoBehaviour
-{
-    public Text finalScoreText;
-
-    private void Start()
-    {
-        if (GameManager.Instance != null)
-        {
-            finalScoreText.text = "Final Score: " + GameManager.Instance.score;
-        }
-    }
-}
