+++ Bee Balloon Game/Assets/Scripts/StartEndScreens/RestartButton.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class RestartButton : MonoBehaviour
+{
+    //merged with end screen manager .cs
+}
